Update completed session modal feedback and upcoming sessions
